Fix a compiler warning with unsigned comparison
git-svn-id: svn://tug.org/texlive/trunk/Build/source@34759 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/dvipdfm-x/tt_cmap.c
@@ -432,7 +432,7 @@ static USHORT
 lookup_cmap12 (struct cmap12 *map, ULONG cccc)
 {
   USHORT gid = 0;
-  ULONG  i;
+  int i;
 
   i = map->nGroups;
   while (i-- >= 0 &&
